requirements.txt
flask
flask-smorest
python-dotenv
marshmallow
sqlalchemy
flask-sqlalchemy
flask-jwt-extended
passlib
flask-migrate
gunicorn
psycopg2-binary
requests
rq
redis